Split out shell command.
